v 1.1 Added import image button Added IA recognition Added pre-trained ia files selector Major Bug Fixes

+++ save_on_location.py
@@ -3,11 +3,23 @@ import shutil
 import os
 
 def save_on_directory():
-    path = filedialog.asksaveasfile(mode='w', defaultextension=".jpg")
 
-    if path is None: 
-        return
-    img = "./image.jpg"
-    shutil.move(img, path.name)
+    if os.path.exists("temp/image.jpg"):
+        img = "temp/image.jpg"
+        path = filedialog.asksaveasfile(mode='w', defaultextension=".jpg", filetypes=(("JPEG (*.JPG)", "*.jpg"),("All Files", "*.*") ))
+        if path is None: 
+            path.close()
+            return
+        shutil.move(img, path.name)
+    
     path.close()
-    os.remove(img)
+
+def import_image():
+    img = "temp/image.jpg"
+    path = filedialog.askopenfilename()
+    
+    print(path)
+    if path is None: 
+        path.close()
+        return
+    shutil.copy(path, img)
